requirements.txt
fastapi
uvicorn
google-generativeai
google-cloud-firestore
google-auth
python-dotenv
PyPDF2